:construction: Subiendo codigo

+++ AED/docencia-notes-solucion/docencia-notes/src/main/java/com/docencia/domain/model/Room.java
@@ -21,7 +21,7 @@ public class Room {
     private String type;
 
     @Column(name = "price_per_night")
-    private Long pricePerNight;
+    private Float pricePerNight;
 
     @ManyToOne
     @Column(name = "hotel_id")
